Move Program visitor
